[#37]refactor: 남아있던 member 삭제
